Client
Here you can send the messages. In the begining this client will communicate with the server and negociate a key to make a safe channel. The client will have a private and public key and will use the provate to make the signature in the message. Here we can achieve the autencity. Then when we receive a message we'll automatically use the public key to understood if the message have autenticity and integrity.